Cache PackageInfo to avoid costly marshalling
Previous implementation of project generation involved calling
`UnityEditor.PackageManager.PackageInfo.FindForAssetPath` in a hot
path (once or twice for each asset in the asset DB).

This is not efficient because:
- Internal native implementation involves traversing the whole package list
and do a case insensitive comparison
- Marshalling of the UpmPackageInfo to its managed counterpart is costly
and it is currently done once per file within each package.

This changeset introduces a better lookup logic and maintain an index of
PackageInfo that avoids both the costly lookups and the repetitive marshalling.

Signed-off-by: Simon Ferquel <simon.ferquel@unity3d.com>